moderator, report, decisionのマルチテナント対応のコントローラの修正
